requirements.txt
# ── Application dependencies ─────────────────────────────────────────────────
fastapi
uvicorn
jinja2
python-multipart

# ── OpenTelemetry SDK & Exporters (pinned for reproducibility) ───────────────
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0

# ── OpenTelemetry Auto-Instrumentation ───────────────────────────────────────
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-logging==0.48b0
setuptools==80.9.0
opentelemetry-sdk-logs